remove share module
