+++ frontend/src/metabase/visualizations/echarts/graph/treemap/model/formatters.ts
@@ -1,7 +1,9 @@
+import { formatPercent } from "metabase/static-viz/lib/numbers";
 import { formatValue } from "metabase/visualizations/lib/formatting";
 import type { ComputedVisualizationSettings } from "metabase/visualizations/types";
 
-import type { TreemapChartColumns } from "./types";
+import { getTreemapTotal } from "./data";
+import type { TreemapChartColumns, TreemapTree } from "./types";
 
 export interface TreemapFormatters {
   value: (value: number) => string;
@@ -15,4 +17,10 @@ export function getTreemapFormatters(
   return {
     value: (value: number) => String(formatValue(value, columnSettings)),
   };
+}
+
+export function getTreemapPercentOfTotalFormatter(tree: TreemapTree) {
+  const total = getTreemapTotal(tree);
+
+  return (value: number) => formatPercent(total === 0 ? 0 : value / total);
 }
